requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

requests~=2.23.0
setuptools~=47.1.1
pandas~=1.0.4
Scrapy~=2.1.0